configuration/panstamp-i2c/requirements.txt
paho-mqtt
influxdb-client
smbus2
pyserial
python-dotenv
pytz
